HOTFIX 3 (#49)
* przeniesienie testow tam gdzie trzeba

* testy jakos wygladaja

* przeniesienie mateusza tam gdzie trzeba

* przepakietowanie mateusza

* troche zmienienie polityki mapowania + update testow

* jakies tam poprawki
