+++ lekizapteki-backend/api/src/main/java/pl/uw/mim/io/lekizapteki/exceptions/NoSuchDiseaseException.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class NoSuchDiseaseException extends RuntimeException {
 
+  // TODO COS JEST ZLE
   private final static String MESSAGE = "Nieprawidłowy numer EAN";
 
   public NoSuchDiseaseException() {
