--- lekizapteki-backend/app/src/main/java/pl.io.lekizapteki/excel/parser/MapMedicinesToEntities.java
@@ -1,12 +1,0 @@
-package pl.io.lekizapteki.excel.parser;
-
-import java.util.List;
-import pl.io.lekizapteki.repositories.excelDatatypes.Medicine;
-
-public class MapMedicinesToEntities {
-
-  // dla kazdego leku bedzie tworzylo entity w bazie
-  public static void map(List<Medicine> medicines) {
-
-  }
-}
