--- lekizapteki-backend/app/src/main/java/pl.io.lekizapteki/excel/parser/enums/ColumnType.java
@@ -1,13 +1,0 @@
-package pl.io.lekizapteki.excel.parser.enums;
-
-public enum ColumnType {
-  INGREDIENT,
-  NAME_FORM_DOSE,
-  PACKAGE,
-  EAN,
-  SALE_PRICE,
-  RETAIL_PRICE,
-  TOTAL_REFUNDING,
-  CHARGE_FACTOR,
-  REFUND
-}
