--- lekizapteki-backend/app/src/main/java/pl.io.lekizapteki/repositories/columnTypes/MedicinePropertySetter.java
@@ -1,6 +1,0 @@
-package pl.io.lekizapteki.repositories.columnTypes;
-
-public interface MedicinePropertySetter {
-
-  void setMedicineProperty(String value);
-}
