--- lekizapteki-backend/app/src/main/java/pl.io.lekizapteki/repositories/columnTypes/SalePriceSetter.java
@@ -1,14 +1,0 @@
-package pl.io.lekizapteki.repositories.columnTypes;
-
-import lombok.AllArgsConstructor;
-import pl.io.lekizapteki.repositories.excelDatatypes.Medicine;
-
-@AllArgsConstructor
-public class SalePriceSetter implements MedicinePropertySetter {
-
-  private final Medicine medicine;
-
-  public void setMedicineProperty(String value) {
-    medicine.setSalePrice(value);
-  }
-}
