--- lekizapteki-backend/app/src/main/java/pl.io.lekizapteki/repositories/excelDatatypes/Medicine.java
@@ -1,17 +1,0 @@
-package pl.io.lekizapteki.repositories.excelDatatypes;
-
-import lombok.Data;
-
-@Data
-public class Medicine {
-
-  private String ingredient;
-  private String nameAndFormAndDose;
-  private String pack;
-  private String ean;
-  private String salePrice;
-  private String retailPrice;
-  private String totalFunding;
-  private String chargeFactor;
-  private String refund;
-}
