--- lekizapteki-backend/app/src/main/java/pl.io.lekizapteki/services/JSONConverter.java
@@ -1,18 +1,0 @@
-package pl.io.lekizapteki.services;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.AllArgsConstructor;
-import lombok.SneakyThrows;
-import org.springframework.stereotype.Component;
-
-@Component
-@AllArgsConstructor
-public class JSONConverter<T> {
-
-  private final ObjectMapper mapper;
-
-  @SneakyThrows
-  public String convertObjectsToJsonString(T objects) {
-    return mapper.writeValueAsString(objects);
-  }
-}
