+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/controlers/DiseaseController.java
@@ -17,4 +17,5 @@ public class DiseaseController implements DiseaseApi {
   public List<DiseaseDto> getDiseases() {
     return getDiseases.execute();
   }
+
 }
