+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/excel/parser/editors/medicine/setter/MedicinePropertySetter.java
@@ -1,0 +1,6 @@
+package pl.uw.mim.io.lekizapteki.excel.parser.editors.medicine.setter;
+
+public interface MedicinePropertySetter {
+
+  void setMedicineProperty(String value);
+}
