+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/excel/parser/mappers/MedicinesToEntitiesMapper.java
@@ -1,0 +1,12 @@
+package pl.uw.mim.io.lekizapteki.excel.parser.mappers;
+
+import java.util.List;
+import pl.uw.mim.io.lekizapteki.excel.parser.models.Medicine;
+
+public class MedicinesToEntitiesMapper {
+
+  // dla kazdego leku bedzie tworzylo entity w bazie
+  public static void map(List<Medicine> medicines) {
+
+  }
+}
