+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/excel/parser/models/Medicine.java
@@ -1,0 +1,17 @@
+package pl.uw.mim.io.lekizapteki.excel.parser.models;
+
+import lombok.Data;
+
+@Data
+public class Medicine {
+
+  private String ingredient;
+  private String nameAndFormAndDose;
+  private String pack;
+  private String ean;
+  private String salePrice;
+  private String retailPrice;
+  private String totalFunding;
+  private String chargeFactor;
+  private String refund;
+}
