+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/mappers/dto/DiseaseDtoMapper.java
@@ -1,0 +1,17 @@
+package pl.uw.mim.io.lekizapteki.mappers.dto;
+
+import lombok.experimental.UtilityClass;
+import pl.uw.mim.io.lekizapteki.models.disease.DiseaseDto;
+import pl.uw.mim.io.lekizapteki.repositories.entities.DiseaseEntity;
+
+@UtilityClass
+public class DiseaseDtoMapper {
+
+  public DiseaseDto map(DiseaseEntity diseaseEntity) {
+    return DiseaseDto.builder()
+        .id(diseaseEntity.getId())
+        .name(diseaseEntity.getName())
+        .build();
+  }
+
+}
