+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/repositories/DiseaseRepository.java
@@ -1,9 +1,11 @@
 package pl.uw.mim.io.lekizapteki.repositories;
 
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.uw.mim.io.lekizapteki.repositories.entities.DiseaseEntity;
 
 public interface DiseaseRepository extends JpaRepository<DiseaseEntity, Long> {
 
-  DiseaseEntity getDiseaseEntityById(Long id);
+  Optional<DiseaseEntity> getDiseaseEntityById(Long id);
+
 }
