+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/services/DiseaseService.java
@@ -15,4 +15,5 @@ public class DiseaseService {
   public List<DiseaseEntity> getAllDiseases() {
     return diseaseRepository.findAll();
   }
+
 }
