+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/usecases/GetDiseases.java
@@ -1,9 +1,10 @@
 package pl.uw.mim.io.lekizapteki.usecases;
 
 import java.util.List;
+import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.uw.mim.io.lekizapteki.mappers.dto.DiseasesDtoMapper;
+import pl.uw.mim.io.lekizapteki.mappers.dto.DiseaseDtoMapper;
 import pl.uw.mim.io.lekizapteki.models.disease.DiseaseDto;
 import pl.uw.mim.io.lekizapteki.services.DiseaseService;
 
@@ -14,6 +15,10 @@ public class GetDiseases {
   private DiseaseService diseaseService;
 
   public List<DiseaseDto> execute() {
-    return DiseasesDtoMapper.map(diseaseService.getAllDiseases());
+    return diseaseService
+        .getAllDiseases().stream()
+        .map(DiseaseDtoMapper::map)
+        .collect(Collectors.toList());
   }
+
 }
