+++ lekizapteki-backend/app/src/main/java/pl/uw/mim/io/lekizapteki/usecases/GetMedicines.java
@@ -1,9 +1,10 @@
 package pl.uw.mim.io.lekizapteki.usecases;
 
 import java.util.List;
+import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.uw.mim.io.lekizapteki.mappers.dto.MedicinesDtoMapper;
+import pl.uw.mim.io.lekizapteki.mappers.dto.MedicineDtoMapper;
 import pl.uw.mim.io.lekizapteki.models.medicine.MedicineDto;
 import pl.uw.mim.io.lekizapteki.services.MedicineService;
 
@@ -14,7 +15,10 @@ public class GetMedicines {
   MedicineService medicineService;
 
   public List<MedicineDto> execute(Long diseaseId) {
-    return MedicinesDtoMapper
-        .map(medicineService.getMedicinesForDiseaseId(diseaseId));
+    return medicineService
+        .getMedicinesForDiseaseId(diseaseId).stream()
+        .map(MedicineDtoMapper::map)
+        .collect(Collectors.toList());
   }
+
 }
