+++ lekizapteki-backend/pom.xml
@@ -93,8 +93,9 @@
 
       <dependency>
         <groupId>org.junit.jupiter</groupId>
-        <artifactId>junit-jupiter</artifactId>
-        <version>5.6.2</version>
+        <artifactId>junit-jupiter-api</artifactId>
+        <version>5.5.2</version>
+        <scope>test</scope>
       </dependency>
 
     </dependencies>
